Add Sensor/Serial combo and .gitignore
Added the new Sensor/Serial Board which, as you would expect,
is a combination of the old boards. In addition, updates were made to
each of the boards circuits. Depricated components were removed from
the sensor board, which essentially reduced it to just the depth sensor
connector, I2C ADC, and level shifter.

The serial board was updated to use the FT232 FTDI chip,
replacing the FT4232. The 4232 is not only overkill for our use case,
but also failed on multiple occasions. These failures were likely due to a
faulty 5V regulator, so the 232's ability to power off the USB's 5V line
(instead of our 5V bus) made it a great replacement. The PCB was updated to
take advantage of this feature, which will hopefully prove more relaiable.

A .gitignore was also added, so that Gerber files (and the like) are
not tracked. These files should be rebuilt every time they are needed,
to ensure that the project outputs match the current board revision.
I meant to include this in a separate commit, but forgot. oh well.